Fix DialogBug
ZenkyuWeb v1.0 has a bug that you cannot record your attendance twice or
more. When you submit your input twice, the latter input is ignored and
the former one is submitted instead of the latter.

In the line 198 of app.js in the commit
933133368924f13c9aa65efed6e2730357088f04,
```
   $(this).dialog("close");
```

This code doesn't removes either values of <#prompt-presence> nor
<#prompt-memo> of first input, should remove both of them, therefore
anything else input is ignored.

In this commit,
```
   $("#prompt").remove()
```
does remove #prompt. Your input is to be recoreded successfully.